[docs] Create language reference manual
I initially thought I'd create most of this document and then ask for
feedback.  But I found that it's difficult to write (for me) without an
implementation to expriment with or to try and use.  Instead I've written
down what I know already and will begin work on the implementation, filling
out this document as I go.

docs/plasma_ref.txt:
    New document

Makefile:
docs/index.txt:
    Add new document