[plasmac] Place all output in an output directory.
In order to support multiple kinds of files the compiler can produce (.pz
and diagnostic information for now) the compiler now puts all outputs in an
output directory, which if unspecified is the directory where the input file
was found.

This change now writes the pretty-printed version of the .pz code to a file
rather than standard output.

src/plasmac.m:
    As above.