[refactor] Add a polymorphic attribute to AST statements
This change adds a polymorphic value to statments in the AST.  This is added
in anticipation of how we will need to track variable lifetimes / visability
on different branches.  Currently it is used to store the statement's
context.

src/ast.m:
    As above.

src/ast.resolve.m:
src/ast_to_core.m:
src/parse.m:
    Conform to above change.