[plasmac] Fix a codegen bug for let expressions
Let expressions did not calculate the depth of the "in" part of the
expression correctly.  Causing code within them to access the wrong stack
slots when retrieving variables.

This change adds more extensive comments to the in-memory PZ bytecode.  This
was used to diagnose the problem.

src/core_to_pz.m:
    As above.