[docs] data constructors and variables must have different case
We now require data constructors to have a capital first letter and
variables and functions to have a lower first letter.  This allows a data
constructor to be lexically and syntactically different, which makes it
easier to have different rules about overloading of these symbols, and
probably other things.

docs/plasma_ref.txt:
    As above.