[plasmac] Encode boolean values more "natively"
A relational expression that is True evaluates to 1 during PZ execution, but
the Bool type's tagged value for True was 4 (a the tag 0 combined with the
word 1).  This would cause crashes when pattern matching a Bool that was
created in this way.

This change adds a new encoding for "strict enums" types with no data
constructors with arguments, it encodes them as a simple word with no tag.
For Bools this means they are encoded as 0 and 1 which works easily with
relational expressions and foreign code (in the future).

src/core.types.m:
    As above.

src/core_to_pz.m:
    Handle new constant encoding.