[plasmac] Provide slightly better type errors
For some type constraints we store the context that created that constraint
for use in error messages.

TODO:
    Better contexts.
    Add context to the other literals.

Generally we should improve typechecking, as the literal that "finds" the
type conflict is only part of what happens when a type conflict occurs.
This may require re-writing the type solver.

This change also improves the type solver generally:
    run_literal no-longer uses multiple values for success/failure
    update/didn't update.

    Reasons for failure are now returned to the solve predicate, which
    throws tham as exceptions.

    run_disj is now more readable.

    Smarter use of types in run_literal, update_domain and unify_domains.

    Added some more trace goals.

src/core.typecheck.solve.m:
    As above.

src/core.typecheck.m:
    Provide literals for the constraints that require it.