[plasmac] Refactor code generation
Refactor code generation so that the generation of each expression takes a
/continuation/: something to execute after the current expression.  This
will make it easier to implement switches more senseably.

At the moment there are two continuation types, "return" (the end of the
function) and "execute these instructions" (for tuples and let expressions.

src/core_to_pz.code.m:
    As above.