[plasmac] Fix a typechecking bug for mutual recursion
Typechecking is intended to work an SCC-at-a-time bottom-up.  This is in
order to support type inference.  However it requires more work to support
both SCC-at-a-time and inference.  For now doing this function-at-a-time is
simpler and makes it work for mutually recursive code.

src/core.typecheck.m:
    As above.

tests/valid/types_4.exp:
tests/valid/types_4.p:
    Extend the test to test for mutually recursive types (and functions).