[plasmac] Refactor creation of the type problem
Change how the type problem is build and transformed to make the type
checker easier to debug.  Rather than flatten tne type problem as it is
created, store it in then nested structure it is created in, then flatten
it before solving it and pretty-print both versions in the debugging output.
This allows us to verify that it is being flattened correctly.

src/core.typecheck.sove.m:
    As above.

src/pretty_utils.m:
    Add some more constant strings.

tests/invalid/types_6.exp:
    Fixup exptected test output.