[plasmac/core] Improve typechecker termination check
A Var/Func literal desn't need to delay the typechecker in every case.
Since resource information only moves in one direction, we only need to
delay based on the groudness of the function arguments, not the resources
wich only need to execute once ever.

This is not the same as when a function type is used within another type,
that code is elsewhere so removing this delay is trivial.

src/core.type_chk.solve.m:
    As above.