[plasmac/core] Use structual equality to compare resources in the type solver
Although resource sets are always potentially incomplete, and therefore
unknown_resources is the same as the empty set, we use the former to detect
problems in the type checker.  So use structual equality here to decide if
the resource set should be replaced and the domain updated.

src/core.type_chk.solve.m:
    As above.