[pzrun] Fix pointer tagging on 64bit systems
Pointer tagging on 64-bit systems mistakenly loaded a 32bit value and
zero-extended it to 64-bit when it should have been sign-extended.  (Because
it was a mask intended to preserve all but the few lowest tag bits).
Now we sign-extend it at runtime.

runtime/pz_builtin.c:
    Sign-extend the value during execution, since this is not supported
    during load-time.

runtime/pz_run.h:
    Add a comment saying that pz_write_instr will always zero-extend
    immediate values.