[plasmac/core] Add a new simplify pass
The simplify pass simplifies the core representation.  This change is going
to make it easier for later stages to find calls in tail position.  At the
moment it removes tuple expressions containing a single expression, and
removes let expressions with either an empty "let" or "in" part (the later
is only valid if the set of bound variables is also empty, otherwise it
would introduce arity errors).

A simple addition that would simplify code a lot is to remove let
expressions that alias an already existing variable.

This pass is run after the semantic checks, so that the checks run on code
that is more-likely to make sense to the developer WRT variable names.

src/core.simplify.m:
    Add new module containing new simplification pass.

src/core.m:
    Include new module.

src/plasmac.m:
    Call the simplification pass.