[pzrun] Allow initialising modules without number parameters
We used to reserve all the storage for module's data when the module was
created.  That's not necessary so we can create a default constructor that
doesn't do that and use it in pz_builtin.cpp, which no-longer needs to
number its closures.

runtime/pz_builtin.cpp:
runtime/pz_module.cpp:
runtime/pz_module.h:
    As above.