[pzrun] Move C++ Future things into a new header file
We follow the practice when we find we want a library feature from a newer
version of a core library we write something similar and place it in a
"future" file.  Then when we update our dependencies to a new version of the
library, say C++17, we check this file and remove the items we no-longer
need.

runtime/pz_cxx_future.h:
    Move future code here.

runtime/pz_util.h:
    From here.

runtime/io_utils.h:
runtime/pz_generic.cpp:
runtime/pz_radix_tree.h:
    Update includes.

docs/C_style.txt:
    Add a section on future modules files such as this.