[pzrun] Allow memory allocation with a NULL top-of-stack
The GC needs the top-of-stack pointer so it can trace the stack and
therefore collect memory.  This hack allows us to allocate memory, but
refuse to collect, when there is a NULL top of stack parameter passed.
This effectively allows allocation at non GC-safe points, which is not
very safe.

runtime/pz_gc.c:
    As above.