[pzrun] Allow the GC to scan a dynamic set of roots
Previously the GC would scan the expression stack to look for heap
pointers.  This change adds a dynamic source of roots, discovered via a
callback which allows other parts of the system to trace for roots, and
provide them to the GC at collection time.

This also makes it possible to run the GC before there is an expression
stack.

runtime/pz_gc.[ch]:
    As above.

runtime/pz.cpp:
    Conform to changes.