[pzrun] Make pz_trace() a static option
PZ_GC_TRACE was a C macro that could be defined to enable tracing of
the GC.  However conditional compilation means that this code is not
checked that it compiles cleanly on every build.  By making it a static
option is is checked, but optimised away when disabled.

runtime/pz_gc.cpp:
    As above.

runtime/pz_option.h:
    Add new options.