[pzrun] Add the builtin module to the program before populating it
Adding the builtin module to the program is the easiest way to have the
GC be able to trace it.  However I generally don't like this order since
in general things could be viewed in an incomplete state if loading
fails or if another thread accesses the structure, neither are a problem
right now though.

runtime/pz.{h,cpp}:
    Add a new_module method.

runtime/pz_builtin.{h,cpp}:
    Change the signature of setup_builtions to accept the module rather
    than return it.

runtime/pz_main.cpp:
    Create a new module then pass it to setup_builtins.