[plasmac/pz] Add code to manage value locations in codegen
During codegen values can be found by the code-being-generated in
various locations, for example as static data linked by the module
closer, or on the stack.  This change starts to add a module that will
manage this information that the code generator will use.

So far this only manages variables and the only place they can be is on
the stack.

src/core_to_pz.locn.m:
src/core_to_pz.m:
    Add new module.

src/core_to_pz.code.m:
    Update code generator.