[plasmac/pz] Value locations can now refer to other locations
To support closures properly a value may be stored in a parent
environment (or grandparent).  So change the val_locn structure so that
it can refer to other val_locns when these things can nest.

Now an environment's struct id is required when the function's captured
set are added to the LocnMap, so provide this and wrap it with the
captured variable list into a new maybe_closure structure.