[compiler/pz] val_locn can now refer to nested locations
To lookup values from a parent environment we need val_locn to be able
to refer to nested locations.  This change modifies the val_locn type
spliting into two and allowing val_locn_next to refer to further
val_locn_next values.  The width of a value is also tracked in val_locn
since it won't always be known for intermediate structure accesses.