[plzrun] Simplify the interface for closures
Resolves #137

The two parts of the runtime, the generic interpreter and the rest have
an arkward interface that was going to make some further changes hard.
This increases the coupling between them but makes the interface
simplier.  To make things modular in the future we will include
different .h files (rather than linking different .o files as we might
in a C project).

runtime/pz_closure.h:
runtime/pz_generic_closure.h:
    As above.

Makefile:
    Add a comment directing people to this file when adding a 2nd
    interpreter.