[plzrun] Use NoGCScope to prevent GCs during object construction
If a GC-able class allocates more GC objects in its constructor, then
it's own memory won't be rooted at that time.  So far we use NoGCScopes
in the callers of these constructors, let's enforce this by using this
type rather than the more general GCCapability in the constructors
themselves.

Do this for Proc and Struct.