[compiler/parse] Check comment tokens to avoid an edge case
An edge case exists with

    /*****/ some_code /* */

The tokeniser does not support non-greedy matching between the beginning
and end of a /* */ comment.  The above will be parsed as a single
comment and some_code will be effectively commented out.  To work around
this search fro */ appearing in the middle of a comment and reject the
program, appologising to the user, if it's found.

Also update the docs and test suite.