[compiler/pre] Self-recursive closures now work
When compiling closures add the closures name to the environment before
recursing into it so that it can find itself.

Note that the compiler generates these as direct calls.  This only works
because a closure has the same environment as itself and no general
attempt at letrecs has been made.