[rt/gc] Make a generic Chunk::contains_pointer() function
ChunkBOP::contains_pointer has been removed and a generic
Chunk::contains_pointer function added.  This function will now return
true for pointers that point into metadata areas of a chunk, but that's
good because it allows us to skip checking other chunks for a pointer
when we know it points into this one (whether it points to a cell or
metadata).