[compiler/pre] Refactor the pre->core transformation
Refactor this transformation so that it happens in-order.

Previously this worked by creating the expression for the tail of a
sequence of statements, then the head and combining those two
expressions into a let.  This was in order to support match statements
that bind one or more variables.

Now it works by computing the expression for the head, then the
expression for the tail and pushing the expression for the tail into the
head's expression.