[compiler/core] Fix indentation in match expressions
When pretty-printing match expressions the whole expression was placed
in a group and therefore the newline before the final } would cause an
indent.  Make the group only part of the match expression and modify
p_parens to take items to place inside and outside the group for correct
indentation when newlines are placed in these groups.