[compiler/core] Refine backtracking in the pretty printer
We need to control backtracking in the pretty printer specifically.
There are two concepts that we now control (but we mix them).

 * Whether a callee may create choice points and explore the search
   space.
 * Whether a caller can handle failure (it has a choice point it may
   retry with).

We combine these because currently we only need 2 of the 4 combinations
due to the way we explore the search space.  If our caller created a
choice point then we do not and instead may backtrack to the caller, or
our caller did not create a choicepoint (or has no further paths to
explore) so we may.