[build] Remove the MCFLAGS2 make variable
We had MCFLAGS and MCFLAGS2 because I didn't want the first one
overriden so easily.  That can be fixed by folding it into MMC_MAKE
because it hardly ever changes and changes about as much as what Mercury
you're using and not build configuration.

MCFLAGS2 is now named MFCLAGS and the first MCFLAGS is in MMC_MAKE.