[compiler] Add different types of pretty groups
The first type of pretty group is for expressions. It's what's been used
so far to format an expression like:

    let x = a +
              b
        y = foo()
        z

The new type is for lists, subsequent items arn't indented by a default
amount, they always start at the beginning, it can look like

    let x = [1,
             2,
             3]

(the group is within the list).