[compiler] Improve pretty_callis
Improve pretty printing of call-ish expressions.  In some conditions the
pretty printer now adds a soft-break to the start of the parameter list.
This allows it to place all parameters in the following line without
adding further breaks, or allow them one-per-line as before but from a
position further to the left.