[compiler/ast] Use q_name to refer to types in the AST
References to types gave a list of qualifiers then a name.  This is
exactly where a q_name should be used.

There's lots of other places in the AST that just use a string; rather
than fixing them all now I'll fix them one at a time as we build the
test cases that check module imports of that kind work.