[compiler] Write out a modules's interface for the --make-interface mode
This change adds a new module to write out an interface file using the
pretty printer.  Compilation begins as usual but after processing all
the declarations (but not definitions) compilation stops so the
interface file can be written.

The change also changes the pretty printer not to print full symbol
names.  We'll revert this once the parser can read the full symbol names
when parsing an interface file.