[plasma/pre] Use a different `env` for processing module interfaces
Module interfaces will have all symbols allways fully qualified.  We
should use a different environment to process them.  It will refer to
the builtin symbols with the builtin module's prefix.

Pretty print the IO resource as builtin.IO so that when it is read in an
interface file it works.