[compiler/parse] Remove "return match" ambiguity
Update the parser and grammar to require return expressions to be the
last expression within a block.  This ensures that if "return" is followed
by "match" or "if" it can never be the beginning of a new statement and
must be a match or if expression.

Update a test output since this changes the parse error reported for
that test.