[compiler/parse] Simplify the var statement
 * Simplify the var statement so that it declares only a single
   variable.
 * Re-order parser priorities to test longer parsers (assignment) before
   shorter ones (var declaration)
 * Update syntax in tests
 * Update some tests expected outputs.

This change ensures that var statement are unambiguously shorter than
assignment statements, and their syntax is clearly different.  The
parser will check if something is an assignment before a var declaration
making it unambiguous - and ensuring that assignments that unpack are
unambiguous from calls since they always have a `=` token.