[compiler/core] Use sets of constructor ids for construction expressions
Currently types may share construction IDs, but so that importing types
between modules can work nicely we need to switch away from this.  Which
means that a symbol may refer to more than one construction IDs.  So the
core and other representations needs to refer to sets of constructor
IDs, which are resolved to single IDs during type checking.