[build] pass the whitelist parameter to compiler
Now that the build tool writes the whitelist and the compiler can check
it.  This patch makes the build tool pass the whitelist parameter to the
compiler.

This patch also adds a test case showing that a module could not be
imported because it was not included in the build file properly.