[rt] Require the foreign code dllib to be specified on the commmand line
The runtime is not supposed to guess about things like this or path
resolution, that'll be a seperate tool in the future.  So for now
require that any foreign code is specified on the command line.  It's
expected in the form `bytecode_file:native_dllib_file` as each native
code file belongs to a bytecode file.