[compiler] Check that the module name in the .p and .build files matches
The compiler has a new --module-name-check flag that allows a caller to
pass the module name it expects is in the module being compiled.  The
plzbuild tool arranges for this to happen so that the compiler will
check that the module name in the .p file matches that in the .build
file.