[compiler] Add better error for bad pragma arguments
The parser will accept any sequence of strings (and later numbers or
other similar literals) for pragma arguments.  But most pragmas will
only understand a specific sequence of arguments.  The foreign_include
pragma accepts a single string.

This creates an error message when the parser accepts the pragma
arguments but the individual pragma does not.